Update -weverything, ready for new test on guac

+++ gnl/classique/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: scarboni <scarboni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 11:14:23 by scarboni          #+#    #+#             */
-/*   Updated: 2020/06/16 21:00:14 by scarboni         ###   ########.fr       */
+/*   Updated: 2020/06/23 13:09:10 by scarboni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,18 +20,18 @@
 typedef struct	s_fd_read_wip
 {
 	int			fd;
-	int			last_ret_read;
-	size_t		size;
+	ssize_t		last_ret_read;
+	ssize_t		size;
 	char		*line_wip;
 }				t_fd_read_wip;
 
-int				ft_strchr(const char *s, int c, size_t *indice);
+int				ft_strchr(const char *s, int c, ssize_t *indice);
 char			*ft_strdup(const char *src);
 size_t			ft_strlcpy(char *dst, const char *src, size_t dstsize);
 size_t			ft_strlen(const char *s);
 int				cut_line_n(char **line, t_fd_read_wip *fd_wip);
 int				append_buffer(t_fd_read_wip *fd_wip, char *buffer,
-int ret_read);
+ssize_t ret_read);
 int				read_full_line(t_fd_read_wip *fd_wip, char **line);
 t_fd_read_wip	*get_current_wip(t_fd_read_wip *current_wip, int fd);
 void			gnl_cleaning(int return_value, t_fd_read_wip **current_wip);
